Resetmychoice command replaced with confstone
