+++ elements/Confstone.mod.json
@@ -1,0 +1,9 @@
+{
+  "_fv": 28,
+  "_type": "command",
+  "definition": {
+    "commandName": "confstone",
+    "permissionLevel": "4",
+    "argsxml": "<xml xmlns=\"https://developers.google.com/blockly/xml\"><block type=\"args_start\" deletable=\"false\" x=\"40\" y=\"40\"><next><block type=\"literal\"><field name=\"name\">rechoice</field><statement name=\"args\"><block type=\"call_procedure\"><field name=\"procedure\">ResetmychoiceProcedure</field></block></statement><next><block type=\"literal\"><field name=\"name\">detransform</field><statement name=\"args\"><block type=\"entity\"><field name=\"name\">player</field><field name=\"type\">players</field><statement name=\"args\"><block type=\"call_procedure\"><field name=\"procedure\">DetransformKeyPriNazhatiiKlavishi</field></block></statement></block></statement><next><block type=\"literal\"><field name=\"name\">book</field><statement name=\"args\"><block type=\"call_procedure\"><field name=\"procedure\">OpenBook</field></block></statement></block></next></block></next></block></next></block></xml>"
+  }
+}
